requirements.txt
docker==6.0.0
GitPython==3.1.29
